공지사항 삭제 오류발견 댓글이 있을시 댓글삭제후 공지사항 삭제되도록 코드 변경

+++ diary/src/main/java/com/example/diary/service/NoticeService.java
@@ -35,9 +35,15 @@ public class NoticeService {
 		  return 0;
 	  }
 	  //공지사항 삭제
-	  public int deleteNotice(Notice notice) {
-		  int row = noticeMapper.deleteNotice(notice);
-		  return row;
+	  public void deleteNotice(Notice notice) {
+		  int row = noticeMapper.selectComment();
+		  if(row != 0) {
+			  noticeMapper.deleteComment(notice);
+			  noticeMapper.deleteNotice(notice);
+		  }
+		  noticeMapper.deleteNotice(notice);
+		   
+		  ;
 	  }
 	  //공지사항 수정
 	  public int updateNotice(Notice notice) {
